probably not gonna use this one unless everyone else likes it
imo icon2_v3 looks better.
part of it is that, try as I might to coax pdn into behaving, I just could not get the brush strokes to look clean in the slightest ;u;
v4 also loses some of the depth, which I feel obfuscates the intent of the icon.